pyproject.toml
[project]
name = "main-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "deepgram-sdk>=6.0.1",
    "edge-tts>=7.2.7",
    "fastapi>=0.135.1",
    "fastapi-mail>=1.6.2",
    "httpx>=0.28.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=4.0.0",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
    "twilio>=9.10.3",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E","W","F","I","B","UP","N","SIM","C4"]
ignore = ["B904", "B008", "E501", "N802", "N806"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"